cleaner file structure for aprs SDK port
